frontend/vite-project/backend/requirements.txt
django
djangorestframework
django-filter
sqlparse
wheel
mysqlclient
django-cors-headers